+++ java/tech/datatype/DoubleReaderIter.java
@@ -5,8 +5,8 @@ import clojure.lang.Keyword;
 
 public class DoubleReaderIter implements IOBase, DoubleIter
 {
-  int idx;
-  int num_elems;
+  long idx;
+  long num_elems;
   DoubleReader reader;
   public DoubleReaderIter(DoubleReader _reader)
   {
@@ -15,7 +15,7 @@ public class DoubleReaderIter implements IOBase, DoubleIter
     reader = _reader;
   }
   public Keyword getDatatype() { return reader.getDatatype(); }
-  public int size() { return num_elems - idx; }
+  public long size() { return num_elems - idx; }
   public boolean hasNext() { return idx < num_elems; }
   public double nextDouble() {
     double retval = reader.read(idx);
